generate/requirements.txt
Genshi==0.5
Jinja2==2.6
Pygments==1.4
Sphinx==1.0.8
argparse==1.1
biplist==0.5
chardet==1.0.1
coverage==3.5.1
docutils==0.8.1
mock==0.7.2
nose==1.1.2
nosexcover==1.0.7
requests==0.6.1
simplejson==2.2.1
-e git+git://github.com/trigger-corp/validictory.git#egg=Validictory
wsgiref==0.1.2
PyHamcrest==1.6
pystache==0.5.3
